+++ packages/kas/src/index.js
@@ -1,2 +1,4 @@
+export {libVersion} from "./version";
+
 export * from "./nodes";
 export {compare} from "./compare";