+++ packages/math-input/src/components/__tests__/integration.test.tsx
@@ -1,18 +1,29 @@
 import "@testing-library/jest-dom";
-import {screen, render, fireEvent, within} from "@testing-library/react";
+import {
+    screen,
+    render,
+    fireEvent,
+    within,
+    waitFor,
+} from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import MathQuill from "mathquill";
 import React, {useState} from "react";
 
+import {KeypadType} from "../../enums";
 import MathInput from "../input/math-input";
-import KeypadContext from "../keypad-context";
+import {KeypadContext, StatefulKeypadContextProvider} from "../keypad-context";
 import KeypadSwitch from "../keypad-switch";
 
-import type {KeypadAPI} from "../../types";
+import type {KeypadConfiguration} from "../../types";
 
 const MQ = MathQuill.getInterface(2);
 
-function InputWithContext() {
+const defaultConfiguration: KeypadConfiguration = {
+    keypadType: KeypadType.FRACTION,
+};
+
+function InputWithContext({keypadConfiguration}) {
     const [value, setValue] = useState<string>("");
 
     return (
@@ -27,7 +38,12 @@ function InputWithContext() {
                             cb();
                         }}
                         onFocus={() => {
-                            keypadElement?.activate();
+                            keypadElement?.configure(
+                                keypadConfiguration,
+                                () => {
+                                    keypadElement?.activate();
+                                },
+                            );
                         }}
                         onBlur={() => {
                             keypadElement?.dismiss();
@@ -56,26 +72,12 @@ function KeypadWithContext() {
     );
 }
 
-function ConnectedMathInput() {
-    const [keypadElement, setKeypadElement] = useState<KeypadAPI | null>();
-    const [renderer, setRenderer] = useState<any>(null);
-    const [scrollableElement, setScrollableElement] =
-        useState<HTMLElement | null>();
-
+function ConnectedMathInput({keypadConfiguration = defaultConfiguration}) {
     return (
-        <KeypadContext.Provider
-            value={{
-                setKeypadElement,
-                keypadElement,
-                setRenderer,
-                renderer,
-                setScrollableElement,
-                scrollableElement,
-            }}
-        >
-            <InputWithContext />
+        <StatefulKeypadContextProvider>
+            <InputWithContext keypadConfiguration={keypadConfiguration} />
             <KeypadWithContext />
-        </KeypadContext.Provider>
+        </StatefulKeypadContextProvider>
     );
 }
 
@@ -107,10 +109,14 @@ describe("math input integration", () => {
 
         fireEvent.touchStart(input);
 
+        await waitFor(() => {
+            expect(screen.getByRole("button", {name: "4"})).toBeVisible();
+        });
+
         expect(screen.getByRole("button", {name: "1"})).toBeVisible();
     });
 
-    it("updates input when using keypad", () => {
+    it("updates input when using keypad", async () => {
         render(<ConnectedMathInput />);
 
         const input = screen.getByLabelText(
@@ -118,6 +124,11 @@ describe("math input integration", () => {
         );
 
         fireEvent.touchStart(input);
+
+        await waitFor(() => {
+            expect(screen.getByRole("button", {name: "4"})).toBeVisible();
+        });
+
         userEvent.click(screen.getByRole("button", {name: "1"}));
 
         // MathQuill is problematic,
@@ -130,7 +141,7 @@ describe("math input integration", () => {
         expect(span1).toBeVisible();
     });
 
-    it("updates input when pressing many numbers", () => {
+    it("updates input when pressing many numbers", async () => {
         render(<ConnectedMathInput />);
 
         const input = screen.getByLabelText(
@@ -138,6 +149,10 @@ describe("math input integration", () => {
         );
 
         fireEvent.touchStart(input);
+
+        await waitFor(() => {
+            expect(screen.getByRole("button", {name: "4"})).toBeVisible();
+        });
 
         const testNumbers = [8, 6, 7, 5, 3, 0, 9];
         testNumbers.forEach((num) => {
@@ -153,7 +168,7 @@ describe("math input integration", () => {
         expect(mathquillInstance.latex()).toBe("8675309");
     });
 
-    it("can handle symbols", () => {
+    it("can handle symbols", async () => {
         render(<ConnectedMathInput />);
 
         const input = screen.getByLabelText(
@@ -161,6 +176,10 @@ describe("math input integration", () => {
         );
 
         fireEvent.touchStart(input);
+
+        await waitFor(() => {
+            expect(screen.getByRole("button", {name: "4"})).toBeVisible();
+        });
 
         userEvent.click(screen.getByRole("button", {name: "4"}));
         userEvent.click(screen.getByRole("button", {name: "2"}));
@@ -173,5 +192,82 @@ describe("math input integration", () => {
             MQ(document.getElementsByClassName("mq-editable-field")[0]);
 
         expect(mathquillInstance.latex()).toBe("42\\%");
+    });
+
+    it("handles fractions correctly in expression", async () => {
+        const keypadConfiguration = {
+            keypadType: KeypadType.EXPRESSION,
+        };
+        render(
+            <ConnectedMathInput keypadConfiguration={keypadConfiguration} />,
+        );
+
+        const input = screen.getByLabelText(
+            "Math input box Tap with one or two fingers to open keyboard",
+        );
+
+        fireEvent.touchStart(input);
+
+        await waitFor(() => {
+            expect(screen.getByRole("button", {name: "4"})).toBeVisible();
+        });
+
+        userEvent.click(screen.getByRole("button", {name: "1"}));
+        userEvent.click(
+            screen.getByRole("button", {
+                name: "Fraction, excluding the current expression",
+            }),
+        );
+        userEvent.click(screen.getByRole("button", {name: "4"}));
+        userEvent.click(
+            screen.getByRole("button", {
+                name: "Navigate right out of the numerator and into the denominator",
+            }),
+        );
+        userEvent.click(screen.getByRole("button", {name: "2"}));
+
+        // MathQuill is problematic,
+        // this is how to get the value of the input directly from MQ
+        const mathquillInstance =
+            // eslint-disable-next-line testing-library/no-node-access
+            MQ(document.getElementsByClassName("mq-editable-field")[0]);
+
+        expect(mathquillInstance.latex()).toBe("1\\frac{4}{2}");
+    });
+
+    it("handles fractions correctly in fraction", async () => {
+        render(<ConnectedMathInput />);
+
+        const input = screen.getByLabelText(
+            "Math input box Tap with one or two fingers to open keyboard",
+        );
+
+        fireEvent.touchStart(input);
+
+        await waitFor(() => {
+            expect(screen.getByRole("button", {name: "4"})).toBeVisible();
+        });
+
+        userEvent.click(screen.getByRole("button", {name: "1"}));
+        userEvent.click(
+            screen.getByRole("button", {
+                name: "Fraction, excluding the current expression",
+            }),
+        );
+        userEvent.click(screen.getByRole("button", {name: "4"}));
+        userEvent.click(
+            screen.getByRole("button", {
+                name: "Navigate right out of the numerator and into the denominator",
+            }),
+        );
+        userEvent.click(screen.getByRole("button", {name: "2"}));
+
+        // MathQuill is problematic,
+        // this is how to get the value of the input directly from MQ
+        const mathquillInstance =
+            // eslint-disable-next-line testing-library/no-node-access
+            MQ(document.getElementsByClassName("mq-editable-field")[0]);
+
+        expect(mathquillInstance.latex()).toBe("1\\frac{4}{2}");
     });
 });