+++ packages/perseus-core/src/index.ts
@@ -4,3 +4,8 @@ export type {
     KeypadContextRendererInterface,
     RendererInterface,
 } from "./types";
+
+// Careful, `version.ts` uses this function so it _must_ be imported above it
+export {addLibraryVersionToPerseusDebug} from "./utils/add-library-version-to-perseus-debug";
+
+export {libVersion} from "./version";