+++ packages/perseus/src/__tests__/item-renderer.test.tsx
@@ -1,5 +1,6 @@
+import {StatefulKeypadContextProvider} from "@khanacademy/math-input";
 import {RenderStateRoot} from "@khanacademy/wonder-blocks-core";
-import {fireEvent, render, screen} from "@testing-library/react";
+import {fireEvent, render, screen, waitFor} from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import * as React from "react";
 import "@testing-library/jest-dom"; // Imports custom matchers
@@ -71,23 +72,25 @@ export const renderQuestion = (
     let renderer: ItemRenderer | null = null;
     const {container} = render(
         <RenderStateRoot>
-            <ItemRenderer
-                ref={(node) => (renderer = node)}
-                apiOptions={apiOptions}
-                item={question}
-                problemNum={0}
-                reviewMode={false}
-                savedState=""
-                controlPeripherals={false}
-                dependencies={testDependenciesV2}
-                {...optionalProps}
-            />
-            {/* The ItemRenderer _requires_ two divs: a work area and hints
+            <StatefulKeypadContextProvider>
+                <ItemRenderer
+                    ref={(node) => (renderer = node)}
+                    apiOptions={apiOptions}
+                    item={question}
+                    problemNum={0}
+                    reviewMode={false}
+                    savedState=""
+                    controlPeripherals={false}
+                    dependencies={testDependenciesV2}
+                    {...optionalProps}
+                />
+                {/* The ItemRenderer _requires_ two divs: a work area and hints
                 area. Without both of these, it fails to render anything! */}
-            <div id="workarea" />
-            <div id="hintsarea" />
+                <div id="workarea" />
+                <div id="hintsarea" />
 
-            <Peripherals />
+                <Peripherals />
+            </StatefulKeypadContextProvider>
         </RenderStateRoot>,
     );
     if (!renderer) {
@@ -309,18 +312,21 @@ describe("item renderer", () => {
             );
         });
 
-        it("should activate the keypad when widget with input is focused", () => {
+        it("should activate the keypad when widget with input is focused", async () => {
             // Arrange
             const {renderer} = renderQuestion(itemWithInput, {
                 isMobile: true,
                 customKeypad: true,
+                useV2Keypad: true,
             });
 
             // Act
             renderer.focus();
 
             // Assert
-            expect(screen.getByLabelText("7")).toBeVisible();
+            await waitFor(() => {
+                expect(screen.getByLabelText("7")).toBeVisible();
+            });
         });
 
         it("should provide current and previous focus paths on focus change to, and away from, a single widget", () => {