+++ packages/perseus/src/item-version.ts
@@ -1,10 +1,10 @@
 import allWidgets from "./all-widgets";
-import version from "./version";
+import {itemDataVersion} from "./perseus-version";
 import * as Widgets from "./widgets";
 
 Widgets.registerWidgets(allWidgets);
 
 const ItemVersion = Widgets.getVersionVector();
-ItemVersion["::renderer::"] = version.itemDataVersion;
+ItemVersion["::renderer::"] = itemDataVersion;
 
 export default ItemVersion;