+++ packages/perseus-editor/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/perseus-editor
 
+## 2.7.6
+
+### Patch Changes
+
+-   Updated dependencies [0761377a]
+-   Updated dependencies [14138bb0]
+-   Updated dependencies [14138bb0]
+-   Updated dependencies [332d5d6d]
+    -   @khanacademy/perseus@11.6.0
+
+## 2.7.5
+
+### Patch Changes
+
+-   Updated dependencies [7e2ae0ef]
+-   Updated dependencies [1dc460c7]
+    -   @khanacademy/perseus@11.5.1
+
 ## 2.7.4
 
 ### Patch Changes